playing-cards-backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
drf-spectacular